refactor: take 'command' and 'query' interfaces to root directory
