--- src/main/java/com/librarySpring/librarySpring/Author/interfaces/Command.java
@@ -1,7 +1,0 @@
-package com.librarySpring.librarySpring.Author.interfaces;
-
-import org.springframework.http.ResponseEntity;
-
-public interface Command<I, O> {
-    ResponseEntity<O> execute(I input);
-}
