+++ src/main/java/com/librarySpring/librarySpring/Author/services/UpdateAuthorService.java
@@ -2,11 +2,11 @@ package com.librarySpring.librarySpring.Author.services;
 
 
 import com.librarySpring.librarySpring.Author.interfaces.AuthorRepository;
-import com.librarySpring.librarySpring.Author.interfaces.Command;
 import com.librarySpring.librarySpring.Author.model.Author;
 import com.librarySpring.librarySpring.Author.model.AuthorDTO;
 import com.librarySpring.librarySpring.Author.model.UpdateAuthorCommand;
 import com.librarySpring.librarySpring.Author.validators.AuthorValidator;
+import com.librarySpring.librarySpring.Command;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
