--- src/main/java/com/librarySpring/librarySpring/Book/interfaces/Query.java
@@ -1,7 +1,0 @@
-package com.librarySpring.librarySpring.Book.interfaces;
-
-import org.springframework.http.ResponseEntity;
-
-public interface Query <I, O> {
-    ResponseEntity<O> execute(I input);
-}
