+++ src/main/java/com/librarySpring/librarySpring/Book/services/CreateBookService.java
@@ -6,13 +6,12 @@ import com.librarySpring.librarySpring.Book.interfaces.BookRepository;
 import com.librarySpring.librarySpring.Book.model.Book;
 import com.librarySpring.librarySpring.Book.model.BookDTO;
 import com.librarySpring.librarySpring.Book.validators.BookValidator;
-import com.librarySpring.librarySpring.Person.interfaces.Command;
+import com.librarySpring.librarySpring.Command;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 
 import java.util.HashSet;
-import java.util.Optional;
 import java.util.Set;
 
 @Service
