+++ src/main/java/com/librarySpring/librarySpring/Book/services/UpdateBookService.java
@@ -5,7 +5,7 @@ import com.librarySpring.librarySpring.Book.model.Book;
 import com.librarySpring.librarySpring.Book.model.BookDTO;
 import com.librarySpring.librarySpring.Book.model.UpdateBookCommand;
 import com.librarySpring.librarySpring.Book.validators.BookValidator;
-import com.librarySpring.librarySpring.Person.interfaces.Command;
+import com.librarySpring.librarySpring.Command;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
