--- src/main/java/com/librarySpring/librarySpring/Person/interfaces/Query.java
@@ -1,7 +1,0 @@
-package com.librarySpring.librarySpring.Person.interfaces;
-
-import org.springframework.http.ResponseEntity;
-
-public interface Query <I, O> {
-    ResponseEntity<O> execute(I input);
-}
